+++ src/routes/songsRoutes.ts
@@ -6,11 +6,14 @@ import {addNewSong, deleteSong, setSongAccessType, validateAccessType} from "../
 import {ResponseError, responseOk} from "../utils/response";
 import {isIdNumeric} from "../utils/isIdNumeric";
 import {AccessType} from "@prisma/client";
+import {getUserPermissions} from "../controllers/userController";
+import {checkIfUserHasPermission} from "../controllers/permissionController";
+import {Permissions} from "../utils/permissions";
 
 export const router: Router = Router()
 
 router.route('/')
-    .post(logInvokedEndpoint, validateAccessType, authenticateUser, upload.single('file'),
+    .post(logInvokedEndpoint, validateAccessType, authenticateUser, getUserPermissions, upload.single('file'),
         async (req: Request, res: Response, next: NextFunction) => {
             const file: Express.Multer.File = req.file as Express.Multer.File
             if (!file) {
@@ -19,6 +22,7 @@ router.route('/')
             const accessType: AccessType = req.query.access as AccessType
             const userData = res.locals.user
             try {
+                checkIfUserHasPermission(res.locals.permissions, Permissions.MANAGE_SONGS)
                 await addNewSong(accessType, file, userData.id)
                 res.status(201).json(responseOk(userData))
             } catch (e) {
@@ -27,21 +31,23 @@ router.route('/')
         })
 
 router.route('/:id')
-    .put(logInvokedEndpoint, isIdNumeric, validateAccessType, authenticateUser, async (req: Request, res: Response, next: NextFunction) => {
+    .put(logInvokedEndpoint, isIdNumeric, validateAccessType, authenticateUser, getUserPermissions, async (req: Request, res: Response, next: NextFunction) => {
         const userData = res.locals.user
         const songId = BigInt(req.params.id)
         const accessType: AccessType = req.query.access as AccessType
         try {
+            checkIfUserHasPermission(res.locals.permissions, Permissions.MANAGE_SONGS)
             const updatedSong = await setSongAccessType(songId, accessType, userData.id)
             res.status(200).json(responseOk(userData, updatedSong))
         } catch (e) {
             next(e)
         }
     })
-    .delete(logInvokedEndpoint, isIdNumeric, authenticateUser, async (req: Request, res: Response, next: NextFunction) => {
+    .delete(logInvokedEndpoint, isIdNumeric, authenticateUser, getUserPermissions, async (req: Request, res: Response, next: NextFunction) => {
         const userData = res.locals.user
         const songId = BigInt(req.params.id)
         try {
+            checkIfUserHasPermission(res.locals.permissions, Permissions.MANAGE_SONGS)
             await deleteSong(songId, userData.id)
             res.status(200).json(responseOk(userData))
         } catch (e) {
